增加 canal 的 spring boot 支持
