+++ buession-canal-springboot/src/main/java/com/buession/canal/springboot/autoconfigure/AbstractMqAdapterProperties.java
@@ -22,37 +22,15 @@
  * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
-package com.buession.springboot.canal.autoconfigure;
+package com.buession.canal.springboot.autoconfigure;
+
+import com.buession.canal.springboot.MqBaseInstanceConfiguration;
 
 /**
  * @author Yong.Teng
- * @since 2.3.1
+ * @since 1.0.0
  */
-abstract class AbstractMqAdapterProperties<IC extends AbstractMqAdapterProperties.MqBaseInstanceConfiguration>
+abstract class AbstractMqAdapterProperties<IC extends MqBaseInstanceConfiguration>
 		extends AbstractAdapterProperties<IC> {
-
-	/**
-	 * MQ 实例基本配置
-	 *
-	 * @author Yong.Teng
-	 * @since 0.0.1
-	 */
-	public static abstract class MqBaseInstanceConfiguration extends BaseInstanceConfiguration {
-
-		private boolean flatMessage;
-
-		public MqBaseInstanceConfiguration() {
-			super();
-		}
-
-		public boolean isFlatMessage() {
-			return flatMessage;
-		}
-
-		public void setFlatMessage(boolean flatMessage) {
-			this.flatMessage = flatMessage;
-		}
-
-	}
 
 }
