+++ buession-canal-springboot/src/main/java/com/buession/canal/springboot/autoconfigure/AdapterClientConfiguration.java
@@ -22,7 +22,7 @@
  * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
-package com.buession.springboot.canal.autoconfigure;
+package com.buession.canal.springboot.autoconfigure;
 
 import com.buession.canal.client.adapter.AdapterClient;
 import com.buession.canal.client.adapter.KafkaAdapterClient;
@@ -42,7 +42,7 @@ import java.util.stream.Collectors;
 
 /**
  * @author Yong.Teng
- * @since 2.3.1
+ * @since 1.0.0
  */
 @Configuration(proxyBeanMethods = false)
 @EnableConfigurationProperties(CanalProperties.class)
