+++ buession-canal-springboot/src/main/java/com/buession/canal/springboot/autoconfigure/ThreadPoolConfiguration.java
@@ -22,10 +22,10 @@
  * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
-package com.buession.springboot.canal.autoconfigure;
+package com.buession.canal.springboot.autoconfigure;
 
 import com.buession.canal.core.concurrent.DefaultCanalThreadPoolExecutor;
-import com.buession.springboot.canal.ThreadConfig;
+import com.buession.canal.springboot.ThreadConfig;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -36,7 +36,7 @@ import java.util.concurrent.ThreadPoolExecutor;
 
 /**
  * @author Yong.Teng
- * @since 2.3.1
+ * @since 1.0.0
  */
 @Configuration(proxyBeanMethods = false)
 @EnableConfigurationProperties({CanalProperties.class})
