+++ buession-canal-springboot/src/main/java/com/buession/canal/springboot/package-info.java
@@ -23,8 +23,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  */
 /**
- * 
- *
  * @author Yong.Teng
  * @since 1.0.0
- */package com.buession.canal.springboot;
+ */
+package com.buession.canal.springboot;
