+++ buession-canal-springboot/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.buession.springboot.canal.autoconfigure.ThreadPoolConfiguration, \
-  com.buession.springboot.canal.autoconfigure.AdapterClientConfiguration, \
-  com.buession.springboot.canal.autoconfigure.CanalConfiguration
+  com.buession.canal.springboot.autoconfigure.ThreadPoolConfiguration, \
+  com.buession.canal.springboot.autoconfigure.AdapterClientConfiguration, \
+  com.buession.canal.springboot.autoconfigure.CanalConfiguration
